common: move Result to capa.common from capa.engine
fixes circular import error in capa.features.freeze
